# swagger-runtime.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Clynic API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Dashboard/resumen": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "parameters": [
          {
            "name": "idSucursal",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Clynic.Application.DTOs.Dashboard.DashboardResumenDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Dashboard/citas-por-dia": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "parameters": [
          {
            "name": "fechaDesde",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "fechaHasta",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "periodo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idSucursal",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Clynic.Application.DTOs.Dashboard.DashboardCitasSerieDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Dashboard/operativo": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "parameters": [
          {
            "name": "idSucursal",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Clynic.Application.DTOs.Dashboard.DashboardOperativoDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/HistorialesClinicos/paciente/{idPaciente}": {
      "get": {
        "tags": [
          "HistorialesClinicos"
        ],
        "parameters": [
          {
            "name": "idPaciente",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Clynic.Application.DTOs.HistorialesClinicos.HistorialClinicoResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "HistorialesClinicos"
        ],
        "parameters": [
          {
            "name": "idPaciente",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Clynic.Application.DTOs.HistorialesClinicos.UpsertHistorialClinicoDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Clynic.Application.DTOs.HistorialesClinicos.UpsertHistorialClinicoDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Clynic.Application.DTOs.HistorialesClinicos.UpsertHistorialClinicoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Clynic.Application.DTOs.HistorialesClinicos.HistorialClinicoResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Clynic.Application.DTOs.Dashboard.DashboardCitaPendienteDto": {
        "type": "object",
        "properties": {
          "idCita": {
            "type": "integer",
            "format": "int32"
          },
          "idSucursal": {
            "type": "integer",
            "format": "int32"
          },
          "nombrePaciente": {
            "type": "string",
            "nullable": true
          },
          "fechaHoraInicioPlan": {
            "type": "string",
            "format": "date-time"
          },
          "fechaHoraFinPlan": {
            "type": "string",
            "format": "date-time"
          },
          "estado": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Clynic.Application.DTOs.Dashboard.DashboardCitasPorDiaDto": {
        "type": "object",
        "properties": {
          "fecha": {
            "type": "string",
            "format": "date-time"
          },
          "totalCitas": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "Clynic.Application.DTOs.Dashboard.DashboardCitasSerieDto": {
        "type": "object",
        "properties": {
          "fechaMinima": {
            "type": "string",
            "format": "date-time"
          },
          "fechaDesde": {
            "type": "string",
            "format": "date-time"
          },
          "fechaHasta": {
            "type": "string",
            "format": "date-time"
          },
          "totalPeriodo": {
            "type": "integer",
            "format": "int32"
          },
          "serie": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Clynic.Application.DTOs.Dashboard.DashboardCitasPorDiaDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Clynic.Application.DTOs.Dashboard.DashboardOperativoDto": {
        "type": "object",
        "properties": {
          "totalPacientes": {
            "type": "integer",
            "format": "int32"
          },
          "totalCitasHoy": {
            "type": "integer",
            "format": "int32"
          },
          "totalCitasPendientes": {
            "type": "integer",
            "format": "int32"
          },
          "totalTrabajadores": {
            "type": "integer",
            "format": "int32"
          },
          "filtradoPorSucursal": {
            "type": "boolean"
          },
          "idSucursal": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pendientes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Clynic.Application.DTOs.Dashboard.DashboardCitaPendienteDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Clynic.Application.DTOs.Dashboard.DashboardResumenDto": {
        "type": "object",
        "properties": {
          "totalPacientes": {
            "type": "integer",
            "format": "int32"
          },
          "totalCitasHoy": {
            "type": "integer",
            "format": "int32"
          },
          "totalTrabajadores": {
            "type": "integer",
            "format": "int32"
          },
          "totalSucursales": {
            "type": "integer",
            "format": "int32"
          },
          "fechaCorte": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "Clynic.Application.DTOs.HistorialesClinicos.HistorialClinicoResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idPaciente": {
            "type": "integer",
            "format": "int32"
          },
          "enfermedadesPrevias": {
            "type": "string",
            "nullable": true
          },
          "medicamentosActuales": {
            "type": "string",
            "nullable": true
          },
          "alergias": {
            "type": "string",
            "nullable": true
          },
          "antecedentesFamiliares": {
            "type": "string",
            "nullable": true
          },
          "observaciones": {
            "type": "string",
            "nullable": true
          },
          "fechaCreacion": {
            "type": "string",
            "format": "date-time"
          },
          "fechaActualizacion": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "Clynic.Application.DTOs.HistorialesClinicos.UpsertHistorialClinicoDto": {
        "type": "object",
        "properties": {
          "enfermedadesPrevias": {
            "maxLength": 4000,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "medicamentosActuales": {
            "maxLength": 4000,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "alergias": {
            "maxLength": 4000,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "antecedentesFamiliares": {
            "maxLength": 4000,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "observaciones": {
            "maxLength": 4000,
            "minLength": 0,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Microsoft.AspNetCore.Mvc.ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Ingresa el token JWT con el prefijo 'Bearer '",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}